Move rich-text type to rich-text package
