+++ packages/rich-text/src/types.ts
@@ -1,0 +1,11 @@
+type RichTextFormat = {
+	type: string;
+};
+type RichTextFormatList = Array< RichTextFormat >;
+export type RichTextValue = {
+	text: string;
+	formats: Array< RichTextFormatList >;
+	replacements: Array< RichTextFormat >;
+	start: number | undefined;
+	end: number | undefined;
+};
